requirements.txt
fastapi[all]~=0.115.8
psycopg2~=2.9.10
sqlalchemy~=2.0.38
passlib~=1.7.4
bcrypt~=4.0.1
python-jose[cryptography]~=3.4.0
alembic~=1.14.1
gunicorn~=23.0.0
uvloop~=0.21.0
httpx~=0.28.1
twilio~=9.4.6
azure-ai-documentintelligence~=1.0.0b4
azure-core~=1.32.0
phonenumbers~=9.0.0
boto3~=1.37.13
pillow~=11.1.0



pdf417gen~=0.8.1
faker~=36.1.1
pytest~=8.3.4
requests~=2.32.3
flake8~=7.1.2